requirements.txt
numpy
Pillow
torch
scipy